Add Timestampable for Trick and Comment
