+++ src/AppBundle/Entity/Comment.php
@@ -3,6 +3,7 @@
 namespace AppBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
+use AppBundle\Entity\Traits\TimestampableTrait;
 
 /**
  * Comment
@@ -13,6 +14,8 @@ use Doctrine\ORM\Mapping as ORM;
  */
 class Comment
 {
+    use TimestampableTrait;
+
     /**
      * @var int
      *
@@ -30,20 +33,6 @@ class Comment
     private $content;
 
     /**
-     * @var \DateTime
-     *
-     * @ORM\Column(name="created", type="datetime")
-     */
-    private $created;
-
-    /**
-     * @var \DateTime
-     *
-     * @ORM\Column(name="updated", type="datetime")
-     */
-    private $updated;
-
-    /**
      * @ORM\ManyToOne(targetEntity="AppBundle\Entity\User")
      * @ORM\JoinColumn(nullable=false)
      */
@@ -54,12 +43,6 @@ class Comment
      * @ORM\JoinColumn(nullable=false)
      */
     private $trick;
-
-    public function __construct()
-    {
-        $this->created = new \Datetime();
-        $this->updated = new \Datetime();
-    }
 
     /**
      * Get id
@@ -93,54 +76,6 @@ class Comment
     public function getContent()
     {
         return $this->content;
-    }
-
-    /**
-     * Set created
-     *
-     * @param \DateTime $created
-     *
-     * @return Comment
-     */
-    public function setCreated($created)
-    {
-        $this->created = $created;
-
-        return $this;
-    }
-
-    /**
-     * Get created
-     *
-     * @return \DateTime
-     */
-    public function getCreated()
-    {
-        return $this->created;
-    }
-
-    /**
-     * Set updated
-     *
-     * @param \DateTime $updated
-     *
-     * @return Comment
-     */
-    public function setUpdated($updated)
-    {
-        $this->updated = $updated;
-
-        return $this;
-    }
-
-    /**
-     * Get updated
-     *
-     * @return \DateTime
-     */
-    public function getUpdated()
-    {
-        return $this->updated;
     }
 
     /**
@@ -189,24 +124,5 @@ class Comment
     public function getTrick()
     {
         return $this->trick;
-    }
-
-    /**
-     * Gets triggered only on insert
-     * @ORM\PrePersist
-     */
-    public function onPrePersist()
-    {
-        $this->created = new \DateTime("now");
-        $this->updated = new \DateTime("now");
-    }
-
-    /**
-     * Gets triggered every time on update
-     * @ORM\PreUpdate
-     */
-    public function onPreUpdate()
-    {
-        $this->updated = new \DateTime("now");
     }
 }
