+++ src/AppBundle/Entity/Trick.php
@@ -6,6 +6,7 @@ use Doctrine\ORM\Mapping as ORM;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 use Symfony\Component\Validator\Constraints as Assert;
 use Gedmo\Mapping\Annotation as Gedmo;
+use AppBundle\Entity\Traits\TimestampableTrait;
 
 /**
  * Trick
@@ -20,6 +21,8 @@ use Gedmo\Mapping\Annotation as Gedmo;
  */
 class Trick
 {
+    use TimestampableTrait;
+
     /**
      * @var int
      *
@@ -51,20 +54,6 @@ class Trick
      * @Assert\Valid()
      */
     private $category;
-
-    /**
-     * @var \DateTime
-     *
-     * @ORM\Column(name="created", type="datetime")
-     */
-    private $created;
-
-    /**
-     * @var \DateTime
-     *
-     * @ORM\Column(name="updated", type="datetime")
-     */
-    private $updated;
 
     /**
      * @ORM\OneToMany(targetEntity="AppBundle\Entity\Media", mappedBy="trick")
@@ -139,54 +128,6 @@ class Trick
     public function getDescription()
     {
         return $this->description;
-    }
-
-    /**
-     * Set created
-     *
-     * @param \DateTime $created
-     *
-     * @return Trick
-     */
-    public function setCreated($created)
-    {
-        $this->created = $created;
-
-        return $this;
-    }
-
-    /**
-     * Get created
-     *
-     * @return \DateTime
-     */
-    public function getCreated()
-    {
-        return $this->created;
-    }
-
-    /**
-     * Set updated
-     *
-     * @param \DateTime $updated
-     *
-     * @return Trick
-     */
-    public function setUpdated($updated)
-    {
-        $this->updated = $updated;
-
-        return $this;
-    }
-
-    /**
-     * Get updated
-     *
-     * @return \DateTime
-     */
-    public function getUpdated()
-    {
-        return $this->updated;
     }
 
     /*
@@ -313,25 +254,6 @@ class Trick
     public function getComments()
     {
         return $this->comments;
-    }
-
-    /**
-     * Gets triggered only on insert
-     * @ORM\PrePersist
-     */
-    public function onPrePersist()
-    {
-        $this->created = new \DateTime("now");
-        $this->updated = new \DateTime("now");
-    }
-
-    /**
-     * Gets triggered every time on update
-     * @ORM\PreUpdate
-     */
-    public function onPreUpdate()
-    {
-        $this->updated = new \DateTime("now");
     }
 
     public function getSlug()
